Basic CLI implementation for calling commands
