+++ Pipfile
@@ -11,3 +11,4 @@ pyqt5 = "*"
 matplotlib = "*"
 opencv-python = "*"
 click = "*"
+colorama = "*"
